spread heater_controller.cpp from heater.cpp
